231229 이성주: 유저아이디 리코일로 관리
